refactor: move pointer→PTY forwarding out of Terminal.tsx
The pane owns two mouse consumers: the local grid selection and the app
behind the PTY. The wheel and button routing that talks to the PTY now
lives in use-terminal-pointer-forward.ts; Terminal.tsx keeps the wiring.
