+++ packages/kobe/src/tui-react/panes/terminal/Terminal.tsx
@@ -36,7 +36,7 @@
 import type { EngineTerminalPresentation } from "@/types/terminal-presentation"
 import type { BoxRenderable, TextRenderable } from "@opentui/core"
 import { StyledText } from "@opentui/core"
-import { useEffect, useLayoutEffect, useMemo, useRef, useState } from "react"
+import { useEffect, useLayoutEffect, useMemo, useState } from "react"
 import { ImeCursorRetention } from "../../../tui/panes/terminal/ime-cursor"
 import { type PtyRegistry, getDefaultPtyRegistry } from "../../../tui/panes/terminal/registry"
 import { rowsToStyledText } from "../../../tui/panes/terminal/sgr-to-text-chunk"
@@ -65,6 +65,7 @@ import { useTerminalGeometry } from "./use-terminal-geometry"
 import { useTerminalHostCursor } from "./use-terminal-host-cursor"
 import { useTerminalPty } from "./use-terminal-pty"
 import { useTerminalReset } from "./use-terminal-reset"
+import { useTerminalPointerForward } from "./use-terminal-pointer-forward"
 import { useTerminalSelection } from "./use-terminal-selection"
 
 /* --------------------------------------------------------------------- */
@@ -191,61 +192,9 @@ export function Terminal(props: TerminalProps) {
     setScrollState((current) => moveViewportScroll(current, snapshot.length, bodyRows, lines, snapshotWindow))
   }
 
-  /**
-   * Emulator order for ANY scroll this pane performs — a wheel tick or a
-   * selection drag hanging past an edge. An app that owns its own scrollback
-   * (mouse tracking, or a fullscreen app on the alternate screen) gets wheel
-   * events; only when it wants neither do we move kobe's local viewport.
-   * Engine tabs are why this matters for the drag: Claude Code runs on the
-   * ALTERNATE screen, where there is no local scrollback to move at all, so a
-   * drag held at the edge has to ask the app to scroll, exactly as the wheel
-   * does. `screenX`/`screenY` are absolute pointer coords. Returns true when
-   * the scroll was forwarded — the selection hook then tracks the content
-   * shifts the app's redraws cause under the fixed snapshot rows.
-   */
-  // True between a press the app accepted and its release.
-  const mouseOwnedByApp = useRef(false)
-
-  const scrollFromPointer = (lines: number, screenX: number, screenY: number): boolean => {
-    if (lines === 0) return false
-    const direction = lines < 0 ? "up" : "down"
-    if (pty && !pty.killed && bodyEl) {
-      const col = Math.max(1, screenX - bodyEl.screenX + 1)
-      const row = Math.max(1, screenY - bodyEl.screenY + 1)
-      if (pty.wheel(direction, col, row)) {
-        for (let i = 1; i < Math.abs(lines); i++) pty.wheel(direction, col, row)
-        return true
-      }
-    }
-    scrollBy(lines)
-    return false
-  }
-
-  /**
-   * Emulator order for a button transition, mirroring `scrollFromPointer`:
-   * the app enabled mouse tracking → encode an SGR event and hand the click
-   * to the app (Claude Code's expandable tool rows, vim, less…), and this
-   * pane's grid selection never sees it. Shift bypasses the app the way
-   * iTerm/kitty do, so text can still be selected out of a mouse-aware app.
-   * A press that was forwarded owns its release too — a drag that leaves the
-   * grid must not fall through into a local selection halfway through.
-   */
-  const forwardMouse = (
-    kind: "down" | "up" | "drag",
-    evt: { button: number; x: number; y: number; modifiers?: { shift: boolean; alt: boolean; ctrl: boolean } },
-  ): boolean => {
-    if (kind === "down") mouseOwnedByApp.current = false
-    if (!pty || pty.killed || !bodyEl) return false
-    if (kind !== "down" && !mouseOwnedByApp.current) return false
-    if (evt.modifiers?.shift) return false
-    if (evt.button !== 0 && evt.button !== 1 && evt.button !== 2) return kind !== "down" && mouseOwnedByApp.current
-    const col = Math.max(1, evt.x - bodyEl.screenX + 1)
-    const row = Math.max(1, evt.y - bodyEl.screenY + 1)
-    const forwarded = pty.click(kind, evt.button, col, row, evt.modifiers)
-    if (kind === "down") mouseOwnedByApp.current = forwarded
-    // A drag the app declined (x10/vt200 tracking) is still its gesture.
-    return forwarded || mouseOwnedByApp.current
-  }
+  // Pointer → PTY routing in emulator order (wheel and buttons); the pane
+  // only scrolls its local viewport when the app wants neither.
+  const { scrollFromPointer, forwardMouse } = useTerminalPointerForward({ pty, bodyEl, scrollBy })
 
   /* --------- viewport slicing ---------- */
